requirements.txt
# Real.
dask>=2024.4.1  # Python 3.11.9 breaks previous Dask
distributed>=2024.4.1  # Python 3.11.9 breaks previous Dask
mlflow-cratedb==2.14.1
pydantic<3
salesforce-merlion>=2,<3
sqlalchemy==2.*

boto3

# Development.
# mlflow-cratedb @ git+https://github.com/crate-workbench/mlflow-cratedb.git@main